feat(t-wise): only consider valid interactions
