+++ src/ddnnf/anomalies/t_wise_sampling/sample_merger.rs
@@ -1,6 +1,5 @@
-use crate::ddnnf::anomalies::t_wise_sampling::data_structure::{Config, Sample};
+use super::{Config, Sample, SamplingResult};
 use crate::Ddnnf;
-use rand::prelude::StdRng;
 
 pub mod similarity_merger;
 pub mod zipping_merger;
@@ -8,31 +7,28 @@ pub mod zipping_merger;
 pub(super) trait SampleMerger {
     /// Creates a new sample by merging two samples.
     /// The merging follows the behaviour defined by the merger.
-    fn merge(&self, node_id: usize, left: &Sample, right: &Sample, rng: &mut StdRng) -> Sample;
+    fn merge(&self, node_id: usize, left: &Sample, right: &Sample) -> Sample;
 
     /// Creates a new sample by merging two samples.
     /// The merging follows the behaviour defined by the merger.
     ///
     /// This method only works in place if the used merger actually overrides this method.
     /// The default implementation calls [Self::merge()] and is therefore not in place.
-    fn merge_in_place(
-        &self,
-        node_id: usize,
-        left: Sample,
-        right: &Sample,
-        rng: &mut StdRng,
-    ) -> Sample {
-        self.merge(node_id, &left, right, rng)
+    fn merge_in_place(&self, node_id: usize, left: Sample, right: &Sample) -> Sample {
+        self.merge(node_id, &left, right)
     }
 
     /// Creates a new sample by merging all given samples.
     /// The merging follows the behaviour defined by the merger.
     /// Returns [Sample::empty] if the given slice is empty.
-    fn merge_all(&self, node_id: usize, samples: &[&Sample], rng: &mut StdRng) -> Sample {
+    fn merge_all(&self, node_id: usize, samples: &[&Sample]) -> Sample {
         samples.iter().fold(Sample::default(), |acc, &sample| {
-            self.merge_in_place(node_id, acc, sample, rng)
+            self.merge_in_place(node_id, acc, sample)
         })
     }
+
+    /// Determines whether a set of results short-circuits to a void sample under the assumptions of this sampler.
+    fn is_void(&self, samples: &[&SamplingResult]) -> bool;
 }
 
 /// This is a marker trait that indicates that a [SampleMerger] is for merging the samples
@@ -50,7 +46,7 @@ pub(super) struct DummyAndMerger<'a> {
 }
 
 impl SampleMerger for DummyAndMerger<'_> {
-    fn merge(&self, _node_id: usize, left: &Sample, right: &Sample, _rng: &mut StdRng) -> Sample {
+    fn merge(&self, _node_id: usize, left: &Sample, right: &Sample) -> Sample {
         if left.is_empty() {
             return right.clone();
         } else if right.is_empty() {
@@ -72,6 +68,10 @@ impl SampleMerger for DummyAndMerger<'_> {
 
         sample
     }
+
+    fn is_void(&self, _samples: &[&SamplingResult]) -> bool {
+        false
+    }
 }
 
 impl AndMerger for DummyAndMerger<'_> {}
@@ -81,7 +81,7 @@ impl AndMerger for DummyAndMerger<'_> {}
 pub(super) struct DummyOrMerger {}
 
 impl SampleMerger for DummyOrMerger {
-    fn merge(&self, _node_id: usize, left: &Sample, right: &Sample, _rng: &mut StdRng) -> Sample {
+    fn merge(&self, _node_id: usize, left: &Sample, right: &Sample) -> Sample {
         if left.is_empty() {
             return right.clone();
         } else if right.is_empty() {
@@ -100,6 +100,10 @@ impl SampleMerger for DummyOrMerger {
             .for_each(|config| sample.add_complete(config));
 
         sample
+    }
+
+    fn is_void(&self, _samples: &[&SamplingResult]) -> bool {
+        false
     }
 }
