+++ src/ddnnf/anomalies/t_wise_sampling/sample_merger/zipping_merger.rs
@@ -1,13 +1,12 @@
-use crate::ddnnf::anomalies::t_wise_sampling::covering_strategies::cover_with_caching;
-use crate::ddnnf::anomalies::t_wise_sampling::data_structure::{Config, Sample};
-use crate::ddnnf::anomalies::t_wise_sampling::sample_merger::{AndMerger, SampleMerger};
-use crate::ddnnf::anomalies::t_wise_sampling::sat_wrapper::SatWrapper;
-use crate::ddnnf::anomalies::t_wise_sampling::t_iterator::TInteractionIter;
+use super::super::{
+    covering_strategies::cover_with_caching_twise, t_iterator::TInteractionIter, SamplingResult,
+    SatWrapper,
+};
+use super::{AndMerger, SampleMerger};
+use super::{Config, Sample};
 use crate::Ddnnf;
 use std::cmp::min;
-
-use rand::prelude::{SliceRandom, StdRng};
-
+use std::collections::HashSet;
 use streaming_iterator::StreamingIterator;
 
 #[derive(Debug, Clone)]
@@ -21,71 +20,126 @@ pub struct ZippingMerger<'a> {
 impl AndMerger for ZippingMerger<'_> {}
 
 impl SampleMerger for ZippingMerger<'_> {
-    fn merge(&self, node_id: usize, left: &Sample, right: &Sample, rng: &mut StdRng) -> Sample {
+    fn merge(&self, node_id: usize, left: &Sample, right: &Sample) -> Sample {
         if left.is_empty() {
             return right.clone();
-        } else if right.is_empty() {
+        }
+
+        if right.is_empty() {
             return left.clone();
         }
 
-        let mut sample =
-            ZippingMerger::zip_samples(left, right, self.ddnnf.number_of_variables as usize);
-        /*
-        Iterate over the remaining interactions. Those are all interactions
-        that contain at least one literal of the left and one of the right subgraph.
-         */
-        let mut left_literals: Vec<i32> = left.get_literals().to_vec();
-        let mut right_literals: Vec<i32> = right.get_literals().to_vec();
-        left_literals.shuffle(rng);
-        right_literals.shuffle(rng);
+        let mut sample = Self::zip_samples(left, right, self.ddnnf.number_of_variables as usize);
 
-        debug_assert!(!left_literals.iter().any(|x| *x == 0));
-        debug_assert!(!right_literals.iter().any(|x| *x == 0));
-
-        for k in 1..self.t {
-            // take k literals of the left subgraph and t-k literals of the right subgraph
-            let left_len = min(left_literals.len(), k);
-            let right_len = min(right_literals.len(), self.t - k);
-            //let left_iter = t_wise_over(left_literals, left_len);
-            let mut left_iter = TInteractionIter::new(&left_literals, left_len);
-
-            while let Some(left_part) = left_iter.next() {
-                //let right_iter = t_wise_over(right_literals, right_len);
-                let mut right_iter = TInteractionIter::new(&right_literals, right_len);
-                while let Some(right_part) = right_iter.next() {
-                    let mut interaction = right_part.to_vec();
-                    interaction.extend_from_slice(left_part);
-                    cover_with_caching(
-                        &mut sample,
-                        &interaction,
-                        self.sat_solver,
-                        node_id,
-                        self.ddnnf.number_of_variables as usize,
-                    );
-                }
-            }
+        for interaction in Self::interactions(left, right, self.t).iter() {
+            cover_with_caching_twise(
+                &mut sample,
+                interaction,
+                self.sat_solver,
+                node_id,
+                self.ddnnf.number_of_variables as usize,
+            );
         }
+
         sample
     }
 
-    fn merge_all(&self, node_id: usize, samples: &[&Sample], rng: &mut StdRng) -> Sample {
+    fn merge_all(&self, node_id: usize, samples: &[&Sample]) -> Sample {
         let (singles, mut samples): (Vec<&Sample>, Vec<&Sample>) =
             samples.iter().partition(|sample| sample.len() <= 1);
 
         let single = singles.iter().fold(Sample::default(), |acc, s| {
-            self.merge_in_place(node_id, acc, s, rng)
+            self.merge_in_place(node_id, acc, s)
         });
 
         samples.push(&single);
         samples.sort_unstable();
 
         samples.iter().fold(Sample::default(), |acc, s| {
-            self.merge_in_place(node_id, acc, s, rng)
+            self.merge_in_place(node_id, acc, s)
         })
+    }
+
+    // For an and node, a single void sample will render the resulting sample void.
+    fn is_void(&self, samples: &[&SamplingResult]) -> bool {
+        samples
+            .iter()
+            .any(|result| matches!(result, SamplingResult::Void))
     }
 }
 
 impl ZippingMerger<'_> {
+    /// Generates all t-wise interactions between two samples.
+    fn interactions(left: &Sample, right: &Sample, t: usize) -> HashSet<Vec<i32>> {
+        Self::generate_interactions(
+            Self::generate_self_interactions(left, t),
+            Self::generate_self_interactions(right, t),
+        )
+    }
+
+    /// Generates all interactions between to sets of interactions.
+    ///
+    /// Both sets are expected to have at least an empty set for each length.
+    fn generate_interactions(
+        left: impl Iterator<Item = HashSet<Vec<i32>>>,
+        right: impl DoubleEndedIterator<Item = HashSet<Vec<i32>>>,
+    ) -> HashSet<Vec<i32>> {
+        let mut interactions = HashSet::new();
+
+        // Both sets are sorted from 1 to t-1.
+        // To combine 1 with t-1, 2 with t-2, etc. we zip left with the reverse of right.
+        // This will give us this exact ordering as both are complete with at least an empty config for each length.
+        left.zip(right.rev()).for_each(|(left, right)| {
+            for left in left {
+                for right in &right {
+                    let mut interaction = left.clone();
+                    interaction.extend_from_slice(right);
+                    interactions.insert(interaction);
+                }
+            }
+        });
+
+        interactions
+    }
+
+    /// Generates a set of interactions inside a sample ordered by interactions size.
+    ///
+    /// The first element will be all interactions of size 1, the second of size 2, ...
+    ///
+    /// In case no interaction of any given size is found, an empty set is still present.
+    fn generate_self_interactions(
+        sample: &Sample,
+        t: usize,
+    ) -> impl DoubleEndedIterator<Item = HashSet<Vec<i32>>> {
+        // Extract all decided literals from the existing configs.
+        let configs: Vec<Vec<i32>> = sample
+            .iter()
+            .map(|config| config.get_decided_literals().collect())
+            .collect();
+
+        // From 1 to t ...
+        (1..t)
+            // consider each length once.
+            .map(move |k| {
+                let mut interactions = HashSet::new();
+                // For each config ...
+                configs.iter().for_each(|config| {
+                    // generate all interactions of the current length ...
+                    TInteractionIter::new(config, min(config.len(), k)).for_each(|interaction| {
+                        // and insert them into the set.
+                        interactions.insert(interaction.to_vec());
+                    });
+                });
+
+                // There needs to be at least one (empty) element for the merge to happen.
+                if interactions.is_empty() {
+                    interactions.insert(Vec::new());
+                }
+
+                interactions
+            })
+    }
+
     /// Creates a new sample by zipping the given samples together. It is assumed that the vars
     /// covered by the samples are disjoint. For decomposable AND nodes this is always the case.
     fn zip_samples(left: &Sample, right: &Sample, number_of_variables: usize) -> Sample {
@@ -123,7 +177,6 @@ mod test {
     use crate::parser::build_ddnnf;
 
     use super::*;
-    use rand::SeedableRng;
     use std::collections::HashSet;
 
     #[test]
@@ -162,17 +215,16 @@ mod test {
         let sat_solver = SatWrapper::new(&ddnnf);
 
         let zipping_merger = ZippingMerger {
-            t: 2,
+            t: 3,
             sat_solver: &sat_solver,
             ddnnf: &ddnnf,
         };
 
-        let mut rng = StdRng::seed_from_u64(42);
         let mut left_sample = new_with_literals(HashSet::from([2, 3]), vec![-2, 3]);
         left_sample.add_partial(Config::from(&[3], 4));
         let right_sample = Sample::new_from_configs(vec![Config::from(&[1, 4], 4)]);
 
-        let result = zipping_merger.merge(node, &left_sample, &right_sample, &mut rng);
+        let result = zipping_merger.merge(node, &left_sample, &right_sample);
         let expected = Sample::new_from_configs(vec![Config::from(&[-2, 1, 3, 4], 4)]);
         assert_eq!(result, expected);
     }
